packages/scraper/requirements.txt
playwright==1.44.0
playwright-stealth==1.0.6
tenacity==8.3.0
Pillow==10.3.0
httpx==0.27.0
supabase==2.5.0
boto3==1.34.0          # Cloudflare R2 (S3-compatible)
python-dotenv==1.0.1
structlog==24.1.0
pydantic==2.7.1
